Fix slow pipeline
This fixes the slow stashing and unstashing in the pipeline with this application,
it also saves lots of disk space on my machine for extra bonus points.

Prior to this change this is the size of all the jars built by this application:
```
$ find . -type f -name *.jar  -exec du -h {} \;
 96M	./repository/build/libs/definition-store-repository-1.2.0-SNAPSHOT.jar
1.4G	./integration-tests/build/libs/integration-tests-1.2.0-SNAPSHOT.jar
624M	./excel-importer/build/libs/definition-store-excel-importer-1.2.0-SNAPSHOT.jar
 56K	./gradle/wrapper/gradle-wrapper.jar
 80M	./aat/build/libs/aat-1.2.0-SNAPSHOT.jar
244K	./aat/build/tmp/expandedArchives/org.jacoco.agent-0.8.0.jar_0a1ac833cae91eb2e2152073ffb36feb/jacocoagent.jar
 62M	./app-insights/build/libs/app-insights-1.2.0-SNAPSHOT.jar
176M	./application/build/libs/case-definition-store-api-1.2.0-SNAPSHOT.jar
112M	./build/libs/case-definition-store-api.jar
252K	./domain/build/libs/test-${project.archivesBaseName}-1.2.0-SNAPSHOT-tests.jar
153M	./domain/build/libs/definition-store-domain-1.2.0-SNAPSHOT.jar
307M	./elastic-search-support/build/libs/definition-store-elastic-search-support-1.2.0-SNAPSHOT.jar
681M	./rest-api/build/libs/definition-store-rest-api-1.2.0-SNAPSHOT.jar
```
= 3.6 GB

The pipeline currently stashes all jars in build/libs folders, the pipeline may be able to change back to just `build/libs/*.jar` instead of `**/build/libs/*.jar`but that would need some testing as fees and pay made that change a long time ago for their apps

Some example stash times from previous runs:
5min14s
6m25s
6m15
6m

Unstash times:
3m7
3m12
4m
2m57

This change should save an average of 9 minutes per run.

Size after this change:
```
$ find . -type f -name *.jar  -exec du -h {} \;
148K	./repository/build/libs/definition-store-repository-1.2.0-SNAPSHOT.jar
140K	./excel-importer/build/libs/definition-store-excel-importer-1.2.0-SNAPSHOT.jar
 56K	./gradle/wrapper/gradle-wrapper.jar
4.0K	./app-insights/build/libs/app-insights-1.2.0-SNAPSHOT.jar
 16K	./application/build/libs/case-definition-store-api-1.2.0-SNAPSHOT.jar
108M	./build/libs/case-definition-store-api.jar
252K	./domain/build/libs/test-domain-1.2.0-SNAPSHOT-tests.jar
184K	./domain/build/libs/definition-store-domain-1.2.0-SNAPSHOT.jar
 32K	./elastic-search-support/build/libs/definition-store-elastic-search-support-1.2.0-SNAPSHOT.jar
 20K	./rest-api/build/libs/definition-store-rest-api-1.2.0-SNAPSHOT.jar
```

= 109MB

96% decrease in size needed to transfer by jenkins!

Jars have been disabled in projects that nothing depend on them, and boot repackaging has been disabled everywhere except the root project which packages the jar for the pipeline