Fix a performance issue loading Event Case Fields
When loading EventCaseFields, the SQL generated by Hibernate
 (5.2.16.final, also tested on 5.4) for these fields causes Postgres
 (9.6.15) to select a query plan that performs a full join of
 event_case_field * case_field * (field_type^3) instead of considering
 only those rows that belong to the CaseTypeId being queried.

Hibernate is generating unnecessary left joins in the subselect used to
 filter event case fields for a case type:

(select events0_.id
 from event events0_
 left outer join state -- irrelevant left join has no effect!
 left outer join... -- more irrelevant joins!
 where events0_.case_type_id=2059)

These joins appear to be confusing the Postgres query optimiser and
stopping it pushing down the where clause into the join, which we fix by
 providing the same where clause without the joins.

In my local testing this fix reduced the runtime of this query from
 ~1800ms to ~10ms.

NB. I also deleted unit tests in CaseTypeServiceImplTest since these
tests are not testing anything - they are simply a plastercast of the
code under test and give false assurance, since we need to run the code
against the full Hibernate/DB stack to have any confidence it is
correct.
There is a strong need for better integration tests.