Migrate webhook_timeouts to a column on webhook
Webhooks typically have a handful of timeouts (3-4) which are stored as
integers. Storing these in a separate table carries a lot of overhead
for such small amounts of data per webhook record.

Please see the migration SQL for details. This change can be run without
downtime so long as definitions are not being imported whilst it runs,
which is current policy).