Add EventWebhook entity joining event and webhook
Moves the three webhooks in event into their own table, which improves
CaseTypeEntity loading time from 7 seconds to ~250m in my test DB of
2000 divorce definition imports. Loading time for display/tab-structure
is also improved from 4 seconds to 400ms, although not the original
purpose of this change.

Currently, events can have three webhooks which fire at event start,
 presubmit and post submit. These were previously modelled as three
 columns on the event table.

This was inefficient to fetch; the webhook table had to be
joined three times to event, but the worst impact was that the SQL
generated by Hibernate ran very poorly on Postgres when joining through
the event table.

This change introduces EventWebhook as a joining table between Event
and Webhook, with webhooks now modelled as a Map<> on Event.