Fix duplicate query on /findJurisdictions
This endpoint is frequently called with a list of IDs to retrieve
details on.

The use of Optional.orElse(jurisdictions.getAll()) was causing all
jurisdictions to be loaded on every request - orElse is not lazily
evaluated, it is just a java method chain, so we when IDs were specified
we would load the requested jurisdictions, then load all jurisdictions
only to throw them away.