Update RC with perf improvements  (#475) (#472)
* Performance improvements for CaseType entity loading (#463)

* Use SubSelect fetch mode on Webhook timeouts

Loading webhook timeouts was causing an N+1 query where every webhook
needed a separate database query to loads its timeouts.

With this change we load the timeouts in a batch reducing total DB
queries from 163 to 43.

* Add index on case_field(case_type_id)

Loading caseFieldAclEntities is taking upwards of 100ms per query
due to a sequential scan on case_field.

Adding an index on case_type_id allows postgres to use an index scan
and reduces this query to < 5ms on my test DB.

* Add .1 suffix to migration 6246

* Add index on webhook_timeouts(webhook_id)

Loading timeouts for a webhook is currently resulting in a sequential
scan of the webhook_timeouts table which we can avoid with this index.

* Create index on event(case_type_id)

This column is frequently queried while loading CaseTypeEntities and is
currently resulting in sequential scans on this large table.

* Fix a performance issue loading Event Case Fields

When loading EventCaseFields, the SQL generated by Hibernate
 (5.2.16.final, also tested on 5.4) for these fields causes Postgres
 (9.6.15) to select a query plan that performs a full join of
 event_case_field * case_field * (field_type^3) instead of considering
 only those rows that belong to the CaseTypeId being queried.

Hibernate is generating unnecessary left joins in the subselect used to
 filter event case fields for a case type:

(select events0_.id
 from event events0_
 left outer join state -- irrelevant left join has no effect!
 left outer join... -- more irrelevant joins!
 where events0_.case_type_id=2059)

These joins appear to be confusing the Postgres query optimiser and
stopping it pushing down the where clause into the join, which we fix by
 providing the same where clause without the joins.

In my local testing this fix reduced the runtime of this query from
 ~1800ms to ~10ms.

NB. I also deleted unit tests in CaseTypeServiceImplTest since these
tests are not testing anything - they are simply a plastercast of the
code under test and give false assurance, since we need to run the code
against the full Hibernate/DB stack to have any confidence it is
correct.
There is a strong need for better integration tests.

* Document EventEntity caseTypeId filter

* Revert "Create index on event(case_type_id)"

This reverts commit 4d8e805a548b651a9506572b8405019a55d7872b.

Duplicate index.

* Migrate webhook_timeouts to a column on webhook

Webhooks typically have a handful of timeouts (3-4) which are stored as
integers. Storing these in a separate table carries a lot of overhead
for such small amounts of data per webhook record.

Please see the migration SQL for details. This change can be run without
downtime so long as definitions are not being imported whilst it runs,
which is current policy).

* Add EventWebhook entity joining event and webhook

Moves the three webhooks in event into their own table, which improves
CaseTypeEntity loading time from 7 seconds to ~250m in my test DB of
2000 divorce definition imports. Loading time for display/tab-structure
is also improved from 4 seconds to 400ms, although not the original
purpose of this change.

Currently, events can have three webhooks which fire at event start,
 presubmit and post submit. These were previously modelled as three
 columns on the event table.

This was inefficient to fetch; the webhook table had to be
joined three times to event, but the worst impact was that the SQL
generated by Hibernate ran very poorly on Postgres when joining through
the event table.

This change introduces EventWebhook as a joining table between Event
and Webhook, with webhooks now modelled as a Map<> on Event.

* Revert "Fix a performance issue loading Event Case Fields"

This reverts commit 2a40c5a35393ea972e22a88dbae1fe1d3e877d54.

* Revert "Add index on webhook_timeouts(webhook_id)"

This reverts commit aa089218f468a15e990481b4c61a6c447b0197a7.

* Rename EventState enum to EventStage

* Rename EventStage to WebhookType

* Replace all reference to webhook state with type in EventEntity

* Lazy fetch EventCaseField parent relationship

* Performance improvements for /display/tab-structure/{id} (#471)

* Use Subselect fetch mode for DisplayGroupEntity fields

* Add index on display_group(case_type_id)

This can avoid sequential scans of this large table when serving UI
display tabs.