Improve performance of AuthorisationCaseFieldParser
Having profiled definition store during an import, a third of all
execution time was spent in the AuthorisationCaseFieldParser.

The old parseAll method was invoked once per case field, and validated
and grouped the entire AuthorisationCaseField sheet which can run to
thousands of rows.

The new addACLEntities eliminates the duplicated work by taking a list
of casefields and only doing the validation & grouping once.