CCD-3438: Improved handling of Idam Feign exceptions
Changed IdamRepository.getUserInfo() to raise an InvalidTokenException or a ServiceException when catching a FeignException, depending on whether the FeignException is a client error or not.  This follows the exception handling that can be found in the ServiceAuthTokenValidator class of the service auth provider java client.

Added tests to IdamRepositoryTest class to test new exception functionality.  Also tidied up class and replaced deprecated methods.