CCD-3438: Improved handling of Idam Feign exceptions (#1251)
* CCD-3438: Improved handling of Idam Feign exceptions

Changed IdamRepository.getUserInfo() to raise an InvalidTokenException or a ServiceException when catching a FeignException, depending on whether the FeignException is a client error or not.  This follows the exception handling that can be found in the ServiceAuthTokenValidator class of the service auth provider java client.

Added tests to IdamRepositoryTest class to test new exception functionality.  Also tidied up class and replaced deprecated methods.

* CCD-3438: Updates following review

Changed IdamRepository.getUserInfo() to log exception message rather than whole exception.  Updated IdamRepositoryTest.errorMessageLogged() in line with this change.

Co-authored-by: Hemanth Potipati <hrpotipati@gmail.com>
Co-authored-by: James Collett <89016121+JamesCollettCGI@users.noreply.github.com>